moves alerting metrics to their own pkg
