+++ pkg/services/ngalert/api/generated_base_api_ruler.go
@@ -17,6 +17,7 @@ import (
 	"github.com/grafana/grafana/pkg/middleware"
 	"github.com/grafana/grafana/pkg/models"
 	apimodels "github.com/grafana/grafana/pkg/services/ngalert/api/tooling/definitions"
+	"github.com/grafana/grafana/pkg/services/ngalert/metrics"
 )
 
 type RulerApiService interface {
@@ -28,61 +29,61 @@ type RulerApiService interface {
 	RoutePostNameRulesConfig(*models.ReqContext, apimodels.PostableRuleGroupConfig) response.Response
 }
 
-func (api *API) RegisterRulerApiEndpoints(srv RulerApiService, metrics *Metrics) {
+func (api *API) RegisterRulerApiEndpoints(srv RulerApiService, m *metrics.Metrics) {
 	api.RouteRegister.Group("", func(group routing.RouteRegister) {
 		group.Delete(
 			toMacaronPath("/api/ruler/{Recipient}/api/v1/rules/{Namespace}"),
-			Instrument(
+			metrics.Instrument(
 				http.MethodDelete,
 				"/api/ruler/{Recipient}/api/v1/rules/{Namespace}",
 				srv.RouteDeleteNamespaceRulesConfig,
-				metrics,
+				m,
 			),
 		)
 		group.Delete(
 			toMacaronPath("/api/ruler/{Recipient}/api/v1/rules/{Namespace}/{Groupname}"),
-			Instrument(
+			metrics.Instrument(
 				http.MethodDelete,
 				"/api/ruler/{Recipient}/api/v1/rules/{Namespace}/{Groupname}",
 				srv.RouteDeleteRuleGroupConfig,
-				metrics,
+				m,
 			),
 		)
 		group.Get(
 			toMacaronPath("/api/ruler/{Recipient}/api/v1/rules/{Namespace}"),
-			Instrument(
+			metrics.Instrument(
 				http.MethodGet,
 				"/api/ruler/{Recipient}/api/v1/rules/{Namespace}",
 				srv.RouteGetNamespaceRulesConfig,
-				metrics,
+				m,
 			),
 		)
 		group.Get(
 			toMacaronPath("/api/ruler/{Recipient}/api/v1/rules/{Namespace}/{Groupname}"),
-			Instrument(
+			metrics.Instrument(
 				http.MethodGet,
 				"/api/ruler/{Recipient}/api/v1/rules/{Namespace}/{Groupname}",
 				srv.RouteGetRulegGroupConfig,
-				metrics,
+				m,
 			),
 		)
 		group.Get(
 			toMacaronPath("/api/ruler/{Recipient}/api/v1/rules"),
-			Instrument(
+			metrics.Instrument(
 				http.MethodGet,
 				"/api/ruler/{Recipient}/api/v1/rules",
 				srv.RouteGetRulesConfig,
-				metrics,
+				m,
 			),
 		)
 		group.Post(
 			toMacaronPath("/api/ruler/{Recipient}/api/v1/rules/{Namespace}"),
 			binding.Bind(apimodels.PostableRuleGroupConfig{}),
-			Instrument(
+			metrics.Instrument(
 				http.MethodPost,
 				"/api/ruler/{Recipient}/api/v1/rules/{Namespace}",
 				srv.RoutePostNameRulesConfig,
-				metrics,
+				m,
 			),
 		)
 	}, middleware.ReqSignedIn)
